Corrige Exercicios 19 ao 24 da lista variaveis-e-tipos-de-dados

+++ exercicios/java-se/primeiros-passos/variaveis-e-tipos-de-dados/pablo-hoties/exercicio20/src/exercicio20/Main.java
@@ -12,9 +12,9 @@ public class Main {
 		double temperaturaEmFahrenheit;
 		
 		System.out.print("Digite uma temperatura em Celsius: ");
-		temperaturaEmCelsius = ler.nextInt();
+		temperaturaEmCelsius = ler.nextDouble();
 		temperaturaEmFahrenheit = 1.8 * temperaturaEmCelsius + 32;
-		System.out.printf("A temperatura em Fahrenheit é: %.0f", temperaturaEmFahrenheit);
+		System.out.printf("A temperatura em Fahrenheit é: %.2f", temperaturaEmFahrenheit);
 		
 	}
